requirements.txt
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
click==6.6
gunicorn==19.6.0
itsdangerous==0.24
requests==2.10.0
google-cloud-vision==0.24.0
